ControlBoardInterfaces.h has been split into multiple new files: IAmplifierControl.h IAxisInfo.h IControlDebug.h IControlLimits.h
