+++ src/libYARP_dev/include/yarp/dev/IControlLimits.h
@@ -1,0 +1,132 @@
+/*
+ * Copyright (C) 2006-2018 Istituto Italiano di Tecnologia (IIT)
+ * Copyright (C) 2006-2010 RobotCub Consortium
+ * All rights reserved.
+ *
+ * This software may be modified and distributed under the terms of the
+ * BSD-3-Clause license. See the accompanying LICENSE file for details.
+ */
+
+#ifndef YARP_DEV_ICONTROLLIMITS_H
+#define YARP_DEV_ICONTROLLIMITS_H
+
+#include <yarp/dev/DeviceDriver.h>
+
+/*! \file IControlLimits.h define control board standard interfaces*/
+
+namespace yarp
+{
+    namespace dev
+    {
+        class IControlLimitsRaw;
+        class IControlLimits;
+    }
+}
+
+/**
+ * @ingroup dev_iface_motor
+ *
+ * Interface for control devices, commands to get/set position and veloity limits.
+ */
+class YARP_dev_API yarp::dev::IControlLimits
+{
+public:
+    /**
+     * Destructor.
+     */
+    virtual ~IControlLimits() {}
+
+    /**
+     * Set the software limits for a particular axis, the behavior of the
+     * control card when these limits are exceeded, depends on the implementation.
+     * @param axis joint number (why am I telling you this)
+     * @param min the value of the lower limit
+     * @param max the value of the upper limit
+     * @return true or false on success or failure
+     */
+    virtual bool setLimits(int axis, double min, double max)=0;
+
+    /**
+     * Get the software limits for a particular axis.
+     * @param axis joint number
+     * @param pointer to store the value of the lower limit
+     * @param pointer to store the value of the upper limit
+     * @return true if everything goes fine, false otherwise.
+     */
+    virtual bool getLimits(int axis, double *min, double *max)=0;
+
+    /**
+     * Set the software speed limits for a particular axis, the behavior of the
+     * control card when these limits are exceeded, depends on the implementation.
+     * @param axis joint number
+     * @param min the value of the lower limit
+     * @param max the value of the upper limit
+     * @return true or false on success or failure
+     */
+    virtual bool setVelLimits(int axis, double min, double max)=0;
+
+    /**
+     * Get the software speed limits for a particular axis.
+     * @param axis joint number
+     * @param min pointer to store the value of the lower limit
+     * @param max pointer to store the value of the upper limit
+     * @return true if everything goes fine, false otherwise.
+     */
+    virtual bool getVelLimits(int axis, double *min, double *max)=0;
+};
+
+/**
+ * Interface for control devices. Limits commands.
+ */
+class YARP_dev_API yarp::dev::IControlLimitsRaw
+{
+public:
+    /**
+     * Destructor.
+     */
+    virtual ~IControlLimitsRaw() {}
+
+    /**
+     *  Set the software limits for a particular axis, the behavior of the
+     * control card when these limits are exceeded, depends on the implementation.
+     * @param axis joint number (why am I telling you this)
+     * @param min the value of the lower limit
+     * @param max the value of the upper limit
+     * @return true or false on success or failure
+     */
+    virtual bool setLimitsRaw(int axis, double min, double max)=0;
+
+    /**
+     * Get the software limits for a particular axis.
+     * @param axis joint number
+     * @param pointer to store the value of the lower limit
+     * @param pointer to store the value of the upper limit
+     * @return true if everything goes fine, false otherwise.
+     */
+    virtual bool getLimitsRaw(int axis, double *min, double *max)=0;
+
+    /**
+     * Set the software speed limits for a particular axis, the behavior of the
+     * control card when these limits are exceeded, depends on the implementation.
+     * @param axis joint number
+     * @param min the value of the lower limit
+     * @param max the value of the upper limit
+     * @return true or false on success or failure
+     */
+    virtual bool setVelLimitsRaw(int axis, double min, double max)=0;
+
+    /**
+     * Get the software speed limits for a particular axis.
+     * @param axis joint number
+     * @param min pointer to store the value of the lower limit
+     * @param max pointer to store the value of the upper limit
+     * @return true if everything goes fine, false otherwise.
+     */
+    virtual bool getVelLimitsRaw(int axis, double *min, double *max)=0;
+};
+
+// interface IControlLimits sets/gets
+constexpr yarp::conf::vocab32_t VOCAB_LIMITS     = yarp::os::createVocab('l','l','i','m');
+constexpr yarp::conf::vocab32_t VOCAB_VEL_LIMITS = yarp::os::createVocab('v','l','i','m');
+
+#endif // YARP_DEV_ICONTROLLIMITS_H
