+++ src/libYARP_dev/include/yarp/dev/ITorqueControl.h
@@ -235,4 +235,18 @@ public:
     virtual bool setMotorTorqueParamsRaw(int j,  const yarp::dev::MotorTorqueParameters params) {return false;}
 };
 
+constexpr yarp::conf::vocab32_t VOCAB_TORQUE            = yarp::os::createVocab('t', 'o', 'r', 'q');
+constexpr yarp::conf::vocab32_t VOCAB_TORQUE_MODE       = yarp::os::createVocab('t', 'r', 'q', 'd');
+constexpr yarp::conf::vocab32_t VOCAB_TRQS              = yarp::os::createVocab('t', 'r', 'q', 's');
+constexpr yarp::conf::vocab32_t VOCAB_TRQ               = yarp::os::createVocab('t', 'r', 'q');
+constexpr yarp::conf::vocab32_t VOCAB_BEMF              = yarp::os::createVocab('b', 'm', 'f');
+constexpr yarp::conf::vocab32_t VOCAB_MOTOR_PARAMS      = yarp::os::createVocab('m', 't', 'p', 's');
+constexpr yarp::conf::vocab32_t VOCAB_RANGES            = yarp::os::createVocab('r', 'n', 'g', 's');
+constexpr yarp::conf::vocab32_t VOCAB_RANGE             = yarp::os::createVocab('r', 'n', 'g');
+constexpr yarp::conf::vocab32_t VOCAB_IMP_PARAM         = yarp::os::createVocab('i', 'p', 'r');
+constexpr yarp::conf::vocab32_t VOCAB_IMP_OFFSET        = yarp::os::createVocab('i', 'o', 'f');
+constexpr yarp::conf::vocab32_t VOCAB_TORQUES_DIRECTS       = yarp::os::createVocab('d', 't', 'q', 's'); //This implements the setRefTorques for the whole part
+constexpr yarp::conf::vocab32_t VOCAB_TORQUES_DIRECT        = yarp::os::createVocab('d', 't', 'q'); //This implements the setRefTorque for a single joint
+constexpr yarp::conf::vocab32_t VOCAB_TORQUES_DIRECT_GROUP  = yarp::os::createVocab('d', 't', 'q', 'g'); //This implements the setRefTorques with joint list
+
 #endif // YARP_DEV_ITORQUECONTROL_H
